Allow more redirects in getNewRedirects
The default max number of redirects for aiohttp is 10. This made the function lose out on a buncha redirects. Add a global variable and set it really high (500) to dictate the max number of redirects allowed. This only affects getNewRedirects and not the domain preprocessing since all domains are looked at manually anyway