Open the tabs in the order they appear in file
Before, it would make a list and then a set of the tabs in `NAME_TABS_TO_OPEN_FILE`. Now it just reads in the domains in the order they appear and then opens them. This will open a tab multiple times if it appears more than once in the file, which is a marginal negative that's vastly outweighed for how I'm using it (which I think is described in the `README` and definitely in the comments at the bottom of to `wikipediaJson.py`) just open them in order.